+++ python/fast_mlsirm/__init__.py
@@ -17,6 +17,7 @@ from .fitstats import (adjusted_chi2_pairs as adjusted_chi2_pairs,
 from .inference import oakes_standard_errors as oakes_standard_errors, observed_information as observed_information, second_order_test as second_order_test, standard_errors_from_vcov as standard_errors_from_vcov, vcov_from_hessian as vcov_from_hessian
 from .linking import link_fixed_item_parameters as link_fixed_item_parameters
 from .linking import irt_link as irt_link, IrtLinkResult as IrtLinkResult
+from .equating import equate_observed_scores as equate_observed_scores, equate_neat as equate_neat, EquateResult as EquateResult
 from .report import render_diagnostics_report as render_diagnostics_report
 from .validation import (ValidationVerdict as ValidationVerdict,
                          validate_judge as validate_judge)
@@ -70,6 +71,9 @@ __all__ = [
     "tcc_drift",
     "irt_link",
     "IrtLinkResult",
+    "equate_observed_scores",
+    "equate_neat",
+    "EquateResult",
     "export_serving_bundle",
     "fit",
     "fit_polytomous",